+++ .github/workflows/integration-tests.yml
@@ -29,10 +29,10 @@ jobs:
             github.com/ChainSafe/gossamer/dot/rpc/modules,
             github.com/ChainSafe/gossamer/dot/rpc/subscription,
             github.com/ChainSafe/gossamer/dot/state,
-            github.com/ChainSafe/gossamer/lib/babe,
             github.com/ChainSafe/gossamer/dot/digest,
             github.com/ChainSafe/gossamer/dot/network,
             github.com/ChainSafe/gossamer/dot/sync,
+            github.com/ChainSafe/gossamer/lib/babe,
             github.com/ChainSafe/gossamer/lib/grandpa,
           ]
     runs-on: ubuntu-latest
